finishing LoginView, WatchView, Repository

+++ data/src/main/java/com/kelvinsugiarto/ministockwatch/data/source/model/CoinInfoResponseEnt.kt
@@ -10,7 +10,7 @@ data class CoinInfoResponseEnt(
     @SerializedName("BlockReward")
     val blockReward: Double,
     @SerializedName("BlockTime")
-    val blockTime: Int,
+    val blockTime: Double,
     @SerializedName("DocumentType")
     val documentType: String,
     @SerializedName("FullName")
